Refactoring Route
